+++ qmcpy/discrete_distribution/korobov/korobov_qrng.py
@@ -55,7 +55,7 @@ class KorobovQRNG(object):
         return x.T
 
     def set_seed(self, seed):
-        self.s = seed if seed else random.randint(2**32)
+        self.s = seed if seed else random.randint(2**32, dtype=uint64)
         return self.s
         
     def set_dimension(self, dimension):
